task_2/requirements.txt
web3==7.9.0
python-dotenv==1.0.1
